==> 01_Primeiros_Passos.R <==
#############################
##  PRIMEIROS PASSOS NO R  ##
#############################


# Cálculos Matemáticos básicos
3 + 4       # soma
10 - 6      # subtração
3 * 5       # multiplicação
20 / 4      # divisão
2**3        # potenciação
sqrt(9)     # raiz quadrada

# Nomeando variáveis
a <- 5
plot(a)
a
b = 8
plot(b)

# Cálculo com variáveis
c <- a * b
c

# Instalação de pacotes
install.packages("dplyr")

# Carregando o pacote
library(dplyr)

# Uso do Help
?plot

# Criando vetores
x = c(2,4,6,8,10,12,14)
y <- c(1,3,5,7,9,11,13)

# Plotando um gráfico
plot(x,y)
